pointsTetrahedra {misc3d}	R Documentation

Create a Set of Tetrahetra Centered at Data Points

Description

Creates a scene consisting of small tetrahedra centered at specified data points in three dimensions.

Usage

pointsTetrahedra(x, y, z, size = 0.01, color = "black", ...)

Arguments

x, y, z
numeric vectors representing point coordinates.

size
numeric; multiple of data range to use for the size of the tetrahedron in each dimension; recycled to length 3.

color
color to use for the tetrahedra.

...
additional arguments to be passed on to makeTriangles.

Details

This function is useful, for example, for incorporating raw data along with a density estimate surface in a scene rendered using standard or grid graphics. For rgl rendering points3d is an alternative.

Value

Returns a triangle mesh scene representing the tetrahedra.

See Also

points3d.

Examples

  with(quakes, {
      d <- kde3d(long, lat, -depth, n = 40)
      v <- contour3d(d$d, exp(-12),d$x/22, d$y/28, d$z/640,
                     color="green", color2="gray", draw=FALSE)
      p <- pointsTetrahedra(long/22, lat/28, -depth/640,
                            size = 0.005)
      drawScene(list(v, p))
  })

[Package misc3d version 0.9-2 Index]