teapot {misc3d}	R Documentation

Utah Teapot

Description

The Utah teapot is a classic computer graphics example. This data set contains a representation in terms of triangles.

Usage

data(teapot)

Format

A list with components vertices and edges. vertices is a 3 by 1976 numeric matrix of the coordinates of the vertices. edges is a 3 by 3751 integer matrix of the indices of the triangles.

Source

Converted from the netCDF file that was at one time made available by Dave Forrest at http://www.maplepark.com/~drf5n/extras/teapot.nc.

[Package misc3d version 0.9-2 Index]